Sphinx: Speedup autoapi computation.

+++ docs/conf.py
@@ -76,13 +76,16 @@ from shutil   import rmtree as shutil_rmtree
 
 if tags.has('PoCCleanUp'):
 	buildDirectory = Path("_build")
-	print("Removing old build directory '{0!s}'...".format(buildDirectory))
-	shutil_rmtree(str(buildDirectory))
+	if (buildDirectory.exists()):
+		print("Removing old build directory '{0!s}'...".format(buildDirectory))
+		shutil_rmtree(str(buildDirectory))
+	else:
+		print("Removing old build directory '{0!s}'... [SKIPPED]".format(buildDirectory))
 
 	pyInfrastructureDirectory = Path("PyInfrastructure")
 	print("Removing created files from '{0!s}'...".format(pyInfrastructureDirectory))
 	for path in pyInfrastructureDirectory.iterdir():
-		if (path.name != "index.rst"):
+		if (path.name.endswith(".rst") and (path.name != (pyInfrastructureDirectory / "index.rst"))):
 			print("  {0!s}".format(path))
 			path.unlink()
 	print()
@@ -92,14 +95,14 @@ autodoc_member_order = "bysource"
 
 # Extract Python documentation and generate ReST files.
 autoapi_modules = {
-  'PoC':        {'output': "PyInfrastructure", 'template': "script"},
-  'Base':       {'output': "PyInfrastructure"},
-  'Compiler':   {'output': "PyInfrastructure"},
-  'DataBase':   {'output': "PyInfrastructure"},
-  'Parser':     {'output': "PyInfrastructure"},
-  'Simulator':  {'output': "PyInfrastructure"},
-  'ToolChain':  {'output': "PyInfrastructure"},
-  'lib':        {'output': "PyInfrastructure"}
+  'PoC':        {'override': False,	'output': "PyInfrastructure", 'template': "script"},
+  'Base':       {'override': False,	'output': "PyInfrastructure"},
+  'Compiler':   {'override': False,	'output': "PyInfrastructure"},
+  'DataBase':   {'override': False,	'output': "PyInfrastructure"},
+  'Parser':     {'override': False,	'output': "PyInfrastructure"},
+  'Simulator':  {'override': False,	'output': "PyInfrastructure"},
+  'ToolChain':  {'override': False,	'output': "PyInfrastructure"},
+  'lib':        {'override': False,	'output': "PyInfrastructure"}
 }
 
 # Add any paths that contain templates here, relative to this directory.
@@ -323,8 +326,8 @@ latex_elements = {
 		% - https://github.com/sphinx-doc/sphinx/issues/3511
 		%
 		\ifdefined\DeclareUnicodeCharacter
-			%\DeclareUnicodeCharacter{2265}{$\geq$}
-			%\DeclareUnicodeCharacter{21D2}{$\Rightarrow$}
+			\DeclareUnicodeCharacter{2265}{$\geq$}
+			\DeclareUnicodeCharacter{21D2}{$\Rightarrow$}
 		\fi
 
 
@@ -498,6 +501,19 @@ changelog_render_changeset = "http://bitbucket.org/myusername/myproject/changese
 # ==============================================================================
 def setup(app):
 	app.add_stylesheet('css/custom.css')
+	if (app.buildername == "pickle"):
+		pyInfrastructureDirectory = Path("PyInfrastructure")
+		print("Removing created files from '{0!s}'...".format(pyInfrastructureDirectory))
+		for path in pyInfrastructureDirectory.iterdir():
+			if (path.name.endswith(".rst") and (path.name != (pyInfrastructureDirectory / "index.rst"))):
+				print("  {0!s}".format(path))
+				path.unlink()
+		print()
+
+		# print(app.config.values['autoapi_modules'])
+		# m = app.config.values['autoapi_modules']
+		# app.config.values['autoapi_modules'] = ({}, m[1], m[2])
+
 	if tags.has('PoCInternal'):
 		app.add_config_value('visibility', 'PoCInternal', True)
 		print("="* 40)
